bug 1133497, Entity reference warnings should ignore all known references.
Entity refs can come from various partial strings, and are generally defined for the
complete l10n context.
Thus it makes sense to refactor the checks to have access to the full reference AST,
and use that to guess known-good entity references.
Refactoring to make it so.
This is tested against a full aurora compare, and only false positives were removed.