bug 1310980, compare-locales parser should reflect full structure of files, r=stas
Rework the existing parsers to reflect the original source structure
better. Explicitly handle comments, and some of the whitespace.

For the defines parser, we also fix some bugs, as that parser
was completely untested, and had bugs.

The Parser.parse() interface doesn't change, clients interested in the
better structure should use Parser.walk()

MozReview-Commit-ID: ARYGW2hrHkw