bug 1315970, failed loading should return empty parse lits, r=flod
When hitting an encoding error, or other IO problems, we should
return an empty parse result.
We used to do that, but regressed that behavior in bug 1290883.

For encoding errors, we could also choose to not error, but
often all strings are returning garbage in that case, not
just the ones with code points triggering the failure.

MozReview-Commit-ID: CAEpIoZBJty