bug 1368449, data storage and iteration, r=mathjazz
This changes the data storage from a dict like
{
 'missingEntity': [
  'onekey',
  'zoom'
 ],
 'obsoleteEntity': [
  'more'
 ]
}
to a list like
[
 {'missingEntity': 'onekey'},
 {'obsoleteEntity': 'more'},
 {'missingEntity': 'zoom'}
]

With this patch there is only a single location where we
sort, which is going to be good for bug 1368444.
I had to rip out the use of SequenceMatcher, because that doesn't
persist order of the keys, it always issues "delete" first, and
then "add", regardless whether the added key would come first
or second. And while I was at that, I dropped the whole
construct of yielding the equal key twice. I did that to mimic
SequenceMatcher, and I don't have to. Makes the code easier
to read, too.

MozReview-Commit-ID: HUWnnBvzyAb