bug 1368025, merge file handling needs to match new configs, r=flod
This change ports the config logic from the old ini-based
approach to the new config-based approach.
Coming along with that are more strict rules about the merge dir,
which will now always have the l10n file structure underneath,
including the locale dir, for firefox apps. For other configs,
it will, for example, create a file
 mergebase/project/locales/file.es-ES.ftl

MozReview-Commit-ID: EuOnNFceVo7