bug 1280894, make AddRemove sort stable without map hacks, r=stas
The logic to have ordering for AddRemove given externally was
a hack, in hindsight.
This patch makes AddRemove stable given two iterables, and adds
tests to verify that.
The test fixes are OK, as the ordering there is for the same
position in the localization, and it's kinda ambiguous which
should come first or not.
This is in support for elmo fluent diffs, 'cause then I
can use AddRemove easier there for both keys and attr names.

MozReview-Commit-ID: 2NEW4lCxW5M