bug 1384901, better handle missing includes, r=flod
This patch does that, for one.

It also has a few forward-looking changes, notably, it allows
for a subclass of TOMLParser to do a different kind of IO.
We'll need this in cross-channel, when we want to load the
configs from a specific version in hg without having to update
the working dir.

MozReview-Commit-ID: EGXPJr83uNu