bug 672031, add check for plural forms, r=flod
Use the list of forms in plurals.py to check if there's an entry
for each of them.
Make those warnings now, as they're likely bugs, but not hard
failures.

The actual `pluralForm` string in intl.properties is also
triggering the plural form logic, so exclude that. Given that we
don't have the file name in the checker, just ignore strings
with that id and a number as value.

MozReview-Commit-ID: E9GkdLredY0