bug 1466836, forbid double straight quotes and check for \' on Android, r=sebastian
We're not checking for stray "-quotes at this time, as they're
used for inline xml extensively. Also, translations have them from
android2po. We're reporting errors for '""', though. Double
quotes just make no sense at all, and make it hard to tell if
a string is considered quoted by Android. Which we need for ...

We're checking that all '-apostrophes are either in a quoted string
or escaped with \.

MozReview-Commit-ID: yrpbFRMZEr