bug 1473577, add validation of reference support, r=stas
We're using None as magic value for validating the reference,
until we're getting to the actual comparison.
Then we're using en-x-moz-reference, which isn't _as_ proof
as None, but is handled better in the result code.

Also adding a plural form for en-x-moz-reference so that
plural checks are run on the reference with English plural
forms.

MozReview-Commit-ID: GvhvFojPbz5