bug 1475969, gettext parser, r=flod
Summary:
Also introduce .localized on Entity.
Now that we have our first bilingual file, we need to start
adding some logic to find out if a bilingual entity is
translated or not.
For our monolingual file formats, that's just True.

I'm using the tuple of msgid and msgctxt as key, we'll need
to get some practical experience to see if that's the right
choice.

Reviewers: flod

Tags: #secure-revision

Bug #: 1475969

Differential Revision: https://phabricator.services.mozilla.com/D3488