bug 1480048, rewrite how project config works to match late, r=stas
This is putting the construction of project config from
its head onto its feet.

We're now generating "smart" paths for entries in toml
configs, which know how to resolve and match for their
complete lifetime.
They're held by Matchers, which keep the environment around
for their lifetime.
If you want to switch the environment of one, create a copy.

This centralized path resolution, and also gets rid of funky
artifacts like "add_global_environment".
And even more so, it removes all the hacks where I expanded
paths depending on whether they were starting with variables
or not, and then some paths had to be called with an environment...

I was young back then.

MozReview-Commit-ID: DbYio3cCZgY