bug 1466088, check for printfs in Android, just %s and %d, r=sebastian
Summary:
Compare reference for each localized parameter.
If there's no reference found, error, as an out-of-bounds parameter
crashes.
This assumes that all parameters are actually used in the Reference,
which should be OK.
If there's a mismatch in the formatter, error.

As in other parts of Android, don't bother about the source
position.

Reviewers: sebastian

Tags: #secure-revision

Bug #: 1466088

Differential Revision: https://phabricator.services.mozilla.com/D3322