bug 1489513, refactor output of compare-locales, r=flod
Preliminaries:

- remove pickling support

This used to be used by elmo to shuffle data from the slave
to the master. We're not doing that no more, removing dead code.

- remove file stats as they're unused.

We never exposed this anywhere, I haven't used it in ages, and
the code is very tied to ini project configs. Making
compare-locales generally useful for other projects would require
as much work as just adding this from scratch.

- remove exhibit output

Exhibit isn't maintained anymore, and I've not been using this
data format in ages.
I used to use it for visual comparisons of different versions
of c-l, but I'm just using the json data these days.

- refactor the commands module before changing arguments

Fixes:
- use paths relative to l10n_base_dir for output
- ignore warnings and errors not for our observer
- switch --quiet to hide obsolete, missing, warnings, errors

The distinction between files and entities isn't as useful.
Let's use --quiet levels for hiding obsolete strings/files,
missing strings/files, warnings, and errors.
Quiet levels now go from 0 to 4. 6 is the same as 4.

- export real json for multiple configs.

This allows to serialize a single json for multiple configs into
a specified file. Think taskcluster artifact.

The written json file is always an array, with the entries
being the <= 3.x json format. This is also true for single
config files to make tooling more consistent.

- reformat text output. Always add missingInFiles and missing

This reformats the summary to be ordered in the way that --quiet
works, errors first, then warnings, missing, obsolete.
Then changed, unchanged, keys.
Then percentage.

It also formats the keys and numbers so they align, for numbers
up to 6 digits.

Also refactored the code to do the details first, for real, instead
of doing them kinda hidden in the return statement.

Reviewers: flod

Tags: #secure-revision

Bug #: 1489513

Differential Revision: https://phabricator.services.mozilla.com/D5517