no bug, separate parser and parsing state
Prior to this change, the parser state was also the parsing state.
So, to add additional state, like to load data from the web,
perhaps in particular revision, one had to add class-global
state my a fixture.
After this change, the parsing state is in a dedicated context,
and the parsing doesn't use any state of the parser. For
included sub context, we create a new context instead of a parser,
and re-use the same parser instance for all parse phases.
Thus parser state can now be used by subclasses to adjust things
like loading.
If a subclass would need to modify the parsing state, it can
subclass ParseContext and TOMLParser.context() to do so.