Path matcher enhancements, rs=foopy
For one, stop the prefix at the first unknown variable expansion.
Then allow to join Matchers, similar to paths.
And use __str__ to expand a Matcher with its own env, and add
a __repr__ to make sure we see what's in a Matcher in shells.
And add __eq__ to allow to compare two Matchers for equality.