no bug, fix ProjectFiles matching for paths w/out vars, rs=foopy
This is mostly bool({}) == False, which I missed earlier.
Now, the fact that this isn't caught in tests is that the tests
actually mocked that very part of the code.
So, much of this patch is actually changing how we test
ProjectFiles, and to only mock the actual IO parts.