Bug 1502009, Entity.pre_comment is not emitted as extra top-level parse item, r=stas
This is a pretty major refactor, in that comments we
associate with a particular Entity are not emitted as
top-level items anymore.
We also absorb the Whitespace between the comment and the
entity.

This brings the legacy parsers in line with what the Fluent
parser did, and allows the merge code to be more strict about
which comment belongs to which entity.

Parse comments first, then white-space.
If an entity follows, create that entity with such pre_comment and
inner white-space. If not, emit comment or white-space as standlone.
It's OK that this might parse whitespace more than once.
Comments are associated with entities if they're not separated by
blank lines. Multiple consecutive comments are joined.

Differential Revision: https://phabricator.services.mozilla.com/D10375