bug 1512709, allow for Entities to force en-US values on serialization, r=stas
Summary:
This allows parsers to highlight entries to be always taken from en-US
instead of localizations.

We use it to add all attributes to the document element in Android's
strings.xml support, notably to add xmlns prefixes that might
be used in elements, and that cross-channel and serializer copy
over to l10n.

Reviewers: stas

Tags: #secure-revision

Bug #: 1512709

Differential Revision: https://phabricator.services.mozilla.com/D13992