bug 1512610, update compare-locales to python-fluent 0.10, r=stas
Summary:
The changes are mostly mechanic about adding back the '-' to
Term IDs.
The other adjustments are around dedention of mutli-line patterns.
It's a bit unclear what the .val should be, so I converted the
tests to .raw_val, which has clear semantics in this context,
i.e., no dedent.

Reviewers: stas

Tags: #secure-revision

Bug #: 1512610

Differential Revision: https://phabricator.services.mozilla.com/D13973