bug 1528627, rewrite Fluent checks to be just in the visitor, r=stas
Summary:
This is a pretty big refactor, and makes the tests more systematic.
Also, I start using message IDs, which should help to move
away from hard-coded messages in the future.

Some changes to the actual tests are "better" wording when it
comes down to messages related to Terms.
Also, check for duplicate attributes on Terms.

Reviewers: stas

Subscribers: flod

Bug #: 1528627

Differential Revision: https://phabricator.services.mozilla.com/D20181