Bug 1542690, support locales in children in project config, r=flod
Previously, locale handling in children and paths in project
config was exclusively filtering.
Now support child configs and paths to explicitly add locales,
and make the top-level l10n.toml just specify the default
for the project and paths.
There's quite a bit of fall-out, and also, the cross-project
multi-locale commandline was just broken. If you ran over
desktop and fennec on a fennec-only locale, it would report
tons of desktop stuff.

Differential Revision: https://phabricator.services.mozilla.com/D26499