Bug 1630599, simplify stats handling, r=stas
The internal stats handling in compare-locales was lazy,
which made it hard to use in other code. That led to errors,
like trying to serialize the word count of obsolete strings,
which doesn't exist. Or trying to ensure a consistent data
profile to other tools like elmo.

I'm also throwing away the difference between missing strings
in completely missing files to missing strings in existing files.
I've not used that in decades, and only worked around this quirk
in other tools or even inside c-l reporting.

Differential Revision: https://phabricator.services.mozilla.com/D71180